requirements.txt
bleach==6.2.0
cassandra-driver==3.29.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dbus-python==1.3.2
fonttools==4.55.2
geomet==0.2.1.post1
idna==3.10
kaggle==1.6.17
kagglehub==0.3.4
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psycopg2-binary==2.9.10
pyparsing==3.2.0
PyQt6==6.7.0
PyQt6_sip==13.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scylla-driver==3.26.9
setuptools==69.5.1.post0
six==1.16.0
text-unidecode==1.3
tqdm==4.67.1
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
